// HappyPaws  Front-End/src/app/models/Albergue.ts
export class Albergue{
    idAlbergue: number = 0;
    nombreAlbergue: string = "";
    direccionAlbergue: string = "";
    telefonoAlbergue: string = "";
    emailAlbergue: string = "";
    capacidadMaxima: number = 0;
    horaApertura: number = 0;
    horaCierre: number = 0;
    webSite: string = "";
    acreditacionAlbergue: string = "";
}